Define env for experiment env in expt script instead of environment.py
