+++ rl_credit/__init__.py
@@ -1,11 +1,3 @@
 from rl_credit.algos import A2CAlgo, PPOAlgo, HCAReturns, HCAState, AttentionAlgo, AttentionQAlgo
 from rl_credit.model import ACModel, RecurrentACModel, ACModelVanilla, ACModelReturnHCA, ACAttention, AttentionQ
 from rl_credit.utils import DictList
-
-
-from gym.envs.registration import register
-
-register(
-    id='GiftDistractorDelay0-v0',
-    entry_point='rl_credit.environment:Delay0_Gifts'
-)
